+++ Editor/BacktraceConfigurationLabels.cs
@@ -7,13 +7,15 @@
         internal static string LABEL_REPORT_PER_MIN = "Reports per minute";
         internal static string LABEL_HANDLE_UNHANDLED_EXCEPTION = "Handle unhandled exceptions";
 
-        internal static string LABEL_IGNORE_SSL_VALIDATION = "Ignore SSL validation";
-        internal static string LABEL_HANDLE_ANR = "Handle ANR (Application not responding)";
-        internal static string LABEL_SEND_UNHANDLED_GAME_CRASHES_ON_STARTUP= "Send unhandled native game crashes on startup";
-        internal static string LABEL_NUMBER_OF_LOGS = "Collect last n game logs";
-        internal static string LABEL_USE_NORMALIZED_EXCEPTION_MESSAGE = "Use normalized exception message";
-        internal static string LABEL_GAME_OBJECT_DEPTH = "Game object depth limit";
         internal static string LABEL_DESTROY_CLIENT_ON_SCENE_LOAD = "Destroy client on new scene load (false - Backtrace managed)";
+        internal static string LABEL_HANDLE_ANR = "Handle ANR (Application not responding)";
+        internal static string LABEL_REPORT_FILTER = "Filter reports";
+        internal static string LABEL_NUMBER_OF_LOGS = "Collect last n game logs";
+        internal static string LABEL_GAME_OBJECT_DEPTH = "Game object depth limit";
+        internal static string LABEL_IGNORE_SSL_VALIDATION = "Ignore SSL validation";
+        internal static string LABEL_SEND_UNHANDLED_GAME_CRASHES_ON_STARTUP= "Send unhandled native game crashes on startup";
+        internal static string LABEL_USE_NORMALIZED_EXCEPTION_MESSAGE = "Use normalized exception message";
+        internal static string LABEL_PERFORMANCE_STATISTICS = "Enable performance statistics";
 
         // database labels
         internal static string LABEL_ENABLE_DATABASE = "Enable Database";
@@ -29,5 +31,7 @@
         internal static string LABEL_RETRY_INTERVAL = "Retry interval";
         internal static string LABEL_RETRY_LIMIT = "Maximum retries";
         internal static string LABEL_RETRY_ORDER = "Retry order (FIFO/LIFO)";
+
+        
     }
 }