+++ Runtime/BacktraceClient.cs
@@ -10,6 +10,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Backtrace.Unity.Tests.Runtime")]
 namespace Backtrace.Unity
 {
     /// <summary>
@@ -140,12 +141,37 @@ namespace Backtrace.Unity
         /// </summary>
         public Func<BacktraceData, BacktraceData> BeforeSend = null;
 
+
+        /// <summary>
+        // Return true to ignore a report, return false to handle the report
+        // and generate one for the error
+        /// </summary>
+        public Func<ReportFilterType, Exception, string, bool> SkipReport = null;
+
         /// <summary>
         /// Set event executed when unhandled application exception event catch exception
         /// </summary>
         public Action<Exception> OnUnhandledApplicationException = null;
 
         private INativeClient _nativeClient;
+
+        public bool EnablePerformanceStatistics
+        {
+            get
+            {
+                return Configuration.PerformanceStatistics;
+            }
+        }
+
+        public int GameObjectDepth
+        {
+            get
+            {
+                return Configuration.GameObjectDepth == 0
+                ? 16 // default maximum game object size
+                : Configuration.GameObjectDepth;
+            }
+        }
 
 
         /// <summary>
@@ -185,8 +211,6 @@ namespace Backtrace.Unity
             }
         }
 
-        private int _gameObjectDepth = 0;
-
         private BacktraceLogManager _backtraceLogManager;
 
         public void OnDisable()
@@ -203,22 +227,21 @@ namespace Backtrace.Unity
 
             Enabled = true;
 
-            // set maximum game object depth
-            _gameObjectDepth = Configuration.GameObjectDepth == 0
-                ? 16 // default maximum game object size
-                : Configuration.GameObjectDepth;
-
             CaptureUnityMessages();
             _reportLimitWatcher = new ReportLimitWatcher(Convert.ToUInt32(Configuration.ReportPerMin));
 
-#if UNITY_2018_4_OR_NEWER
 
             BacktraceApi = new BacktraceApi(
                 credentials: new BacktraceCredentials(Configuration.GetValidServerUrl()),
-                ignoreSslValidation: Configuration.IgnoreSslValidation);
+
+#if UNITY_2018_4_OR_NEWER
+                ignoreSslValidation: Configuration.IgnoreSslValidation
 #else
-            BacktraceApi = new BacktraceApi(new BacktraceCredentials(Configuration.GetValidServerUrl()));
+                ignoreSslValidation: false
 #endif
+                );
+            BacktraceApi.EnablePerformanceStatistics = Configuration.PerformanceStatistics;
+
             if (!Configuration.DestroyOnLoad)
             {
                 DontDestroyOnLoad(gameObject);
@@ -278,6 +301,7 @@ namespace Backtrace.Unity
               attachmentPaths: attachmentPaths,
               attributes: attributes);
             _backtraceLogManager.Enqueue(report);
+
             SendReport(report);
         }
 
@@ -293,6 +317,7 @@ namespace Backtrace.Unity
             {
                 return;
             }
+
             var report = new BacktraceReport(exception, attributes, attachmentPaths);
             _backtraceLogManager.Enqueue(report);
             SendReport(report);
@@ -336,8 +361,19 @@ namespace Backtrace.Unity
         /// <returns>IEnumerator</returns>
         private IEnumerator CollectDataAndSend(BacktraceReport report, Action<BacktraceResult> sendCallback = null)
         {
+            var queryAttributes = new Dictionary<string, string>();
+            var stopWatch = EnablePerformanceStatistics
+                ? System.Diagnostics.Stopwatch.StartNew()
+                : new System.Diagnostics.Stopwatch();
+
             BacktraceData data = SetupBacktraceData(report);
-            yield return new WaitForEndOfFrame();
+
+            if (EnablePerformanceStatistics)
+            {
+                stopWatch.Stop();
+                queryAttributes["performance.report"] = stopWatch.GetMicroseconds();
+            }
+
             if (BeforeSend != null)
             {
                 data = BeforeSend.Invoke(data);
@@ -348,30 +384,51 @@ namespace Backtrace.Unity
             }
             BacktraceDatabaseRecord record = null;
 
-            // avoid serializing data twice
-            // if record is here we should try to send json data that are available in record
-            // otherwise we can still use BacktraceData.ToJson().
-            string json = string.Empty;
             if (Database != null)
             {
                 yield return new WaitForEndOfFrame();
+                if (EnablePerformanceStatistics)
+                {
+                    stopWatch.Restart();
+                }
                 record = Database.Add(data);
                 // handle situation when database refuse to store report.
                 if (record != null)
                 {
                     //Extend backtrace data with additional attachments from backtrace database
                     data = record.BacktraceData;
+                    if (EnablePerformanceStatistics)
+                    {
+                        stopWatch.Stop();
+                        queryAttributes["performance.database"] = stopWatch.GetMicroseconds();
+                    }
+
+
                     if (record.Duplicated)
                     {
                         yield break;
                     }
-                    json = record.BacktraceDataJson();
                 }
             }
-            if (string.IsNullOrEmpty(json))
+
+            yield return new WaitForEndOfFrame();
+            if (EnablePerformanceStatistics)
             {
-                json = data.ToJson();
+                stopWatch.Restart();
             }
+            // avoid serializing data twice
+            // if record is here we should try to send json data that are available in record
+            // otherwise we can still use BacktraceData.ToJson().            
+            string json = record != null
+                ? record.BacktraceDataJson()
+                : data.ToJson();
+
+            if (EnablePerformanceStatistics)
+            {
+                stopWatch.Stop();
+                queryAttributes["performance.json"] = stopWatch.GetMicroseconds();
+            }
+
             //backward compatibility 
             if (RequestHandler != null)
             {
@@ -379,7 +436,12 @@ namespace Backtrace.Unity
                 yield break;
             }
 
-            StartCoroutine(BacktraceApi.Send(json, data.Attachments, data.Deduplication, (BacktraceResult result) =>
+            if (data.Deduplication != 0)
+            {
+                queryAttributes["_mod_duplicate"] = data.Deduplication.ToString();
+            }
+
+            StartCoroutine(BacktraceApi.Send(json, data.Attachments, queryAttributes, (BacktraceResult result) =>
             {
                 if (record != null)
                 {
@@ -410,6 +472,7 @@ namespace Backtrace.Unity
         /// <returns>Backtrace data</returns>
         private BacktraceData SetupBacktraceData(BacktraceReport report)
         {
+
             // apply _mod fingerprint attribute when client should use
             // normalized exception message instead environment stack trace
             // for exceptions without stack trace.
@@ -432,8 +495,7 @@ namespace Backtrace.Unity
                 reportAttributes = _nativeClient.GetAttributes();
             }
 
-            var data = report.ToBacktraceData(reportAttributes, _gameObjectDepth);
-            return data;
+            return report.ToBacktraceData(reportAttributes, GameObjectDepth);
         }
 
 #if UNITY_ANDROID
@@ -445,7 +507,8 @@ namespace Backtrace.Unity
         {
             const string anrMessage = "ANRException: Blocked thread detected";
             _backtraceLogManager.Enqueue(new BacktraceUnityMessage(anrMessage, stackTrace, LogType.Error));
-            SendUnhandledException(new BacktraceUnityMessage(anrMessage, stackTrace, LogType.Error));
+            var hang = new BacktraceUnhandledException(anrMessage, stackTrace);
+            SendUnhandledException(hang);
         }
 #endif
 
@@ -473,13 +536,13 @@ namespace Backtrace.Unity
             _backtraceLogManager.Enqueue(unityMessage);
             if (Configuration.HandleUnhandledExceptions && unityMessage.IsUnhandledException())
             {
-                SendUnhandledException(unityMessage);
+                var exception = new BacktraceUnhandledException(unityMessage.Message, unityMessage.StackTrace);
+                SendUnhandledException(exception);
             }
         }
 
-        private void SendUnhandledException(BacktraceUnityMessage unityMessage)
+        private void SendUnhandledException(BacktraceUnhandledException exception)
         {
-            var exception = new BacktraceUnhandledException(unityMessage.Message, unityMessage.StackTrace);
             if (OnUnhandledApplicationException != null)
             {
                 OnUnhandledApplicationException.Invoke(exception);
@@ -492,6 +555,20 @@ namespace Backtrace.Unity
 
         private bool ShouldSendReport(Exception exception, List<string> attachmentPaths, Dictionary<string, string> attributes)
         {
+            // guess report type
+            var filterType = ReportFilterType.Exception;
+            if (exception is BacktraceUnhandledException)
+            {
+                filterType = (exception as BacktraceUnhandledException).Classifier == "ANRException"
+                    ? ReportFilterType.Hang
+                    : ReportFilterType.UnhandledException;
+            }
+
+
+            if (ShouldSkipReport(filterType, exception, string.Empty))
+            {
+                return false;
+            }
             //check rate limiting
             bool shouldProcess = _reportLimitWatcher.WatchReport(new DateTime().Timestamp());
             if (shouldProcess)
@@ -511,6 +588,11 @@ namespace Backtrace.Unity
 
         private bool ShouldSendReport(string message, List<string> attachmentPaths, Dictionary<string, string> attributes)
         {
+            if (ShouldSkipReport(ReportFilterType.Message, null, message))
+            {
+                return false;
+            }
+
             //check rate limiting
             bool shouldProcess = _reportLimitWatcher.WatchReport(new DateTime().Timestamp());
             if (shouldProcess)
@@ -530,6 +612,15 @@ namespace Backtrace.Unity
 
         private bool ShouldSendReport(BacktraceReport report)
         {
+            if (ShouldSkipReport(
+                    report.ExceptionTypeReport
+                        ? ReportFilterType.Exception
+                        : ReportFilterType.Message,
+                    report.Exception,
+                    report.Message))
+            {
+                return false;
+            }
             //check rate limiting
             bool shouldProcess = _reportLimitWatcher.WatchReport(new DateTime().Timestamp());
             if (shouldProcess)
@@ -572,6 +663,20 @@ namespace Backtrace.Unity
             return !invalidConfiguration;
         }
 
+
+        /// <summary>
+        /// Check if client should skip current report
+        /// </summary>
+        /// <param name="type">Report type</param>
+        /// <param name="exception">Exception object</param>
+        /// <param name="message">String message</param>
+        /// <returns>true if client should skip report. Otherwise false.</returns>
+        private bool ShouldSkipReport(ReportFilterType type, Exception exception, string message)
+        {
+            return Configuration.ReportFilterType.HasFlag(type)
+                || (SkipReport != null && SkipReport.Invoke(type, exception, message));
+
+        }
 
     }
 }