+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Backtrace Unity Release Notes
 
 ## Version 3.0.3
+This release includes significant improvements to performance by way of report filtering as well as improved performance diagnostics. Learn more below.
+ 
+- `BacktraceClient` now supports report filtering. Report filtering is enabled by using the `Filter reports` option in the user interface or for more advanced use-cases, the `SkipReport` delegate available in the BacktraceClient.
+ 
+Sample code: 
+```csharp
+  // Return true to ignore a report, return false to handle the report
+  // and generate one for the error.
+  BacktraceClient.SkipReport = (ReportFilterType type, Exception e, string msg) =>
+  {
+    // ReportFilterType is one of None, Message, Exception,
+    // UnhandledException or Hang. It is also possible to
+    // to filter based on the exception and exception message.
+
+    // Report hangs and crashes only.
+    return type != ReportFilterType.Hang && type != ReportFilterType.UnhandledException;
+  };
+```
+ 
+For example, to only get error reporting for hangs or crashes then only return false for Hang or UnhandledException or set the corresponding options in the user interface as shown below.
+
+![Sample report filter](./Documentation~/images/report-filter.PNG)
+- Support for backtrace-unity timing observability. To enable sending performance information to Backtrace set the`Enable performance statistics` option in the UI. Attributes are created under the performance.* namespace, time unit is microseconds: 
+  * Report creation time (`performance.report`),
+  * JSON serialization time (`performance.json`),
+  * Database add operation time (`performance.database`),
+  * Database single send method time (`performance.send`),
+  * Database single flush method time (`performance.flush`)
 - Improvements to JIT stack frame parsing.
 
 ## Version 3.0.2